add new base card and refactor camera
