+++ web/frontend/package.json
@@ -8,7 +8,7 @@
     "@mdi/js": "^6.9.96",
     "@viamrobotics/prime": "^0.0.106",
     "@viamrobotics/rpc": "^0.1.34",
-    "@viamrobotics/sdk": "^0.0.13",
+    "@viamrobotics/sdk": "^0.0.18",
     "@vitejs/plugin-vue": "^3.1.2",
     "@vueuse/core": "^9.3.1",
     "eslint-import-resolver-custom-alias": "^1.3.0",
