+++ web/frontend/src/components/camera.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { onMounted } from 'vue';
+import { onMounted, onUnmounted, watch } from 'vue';
 import { displayError } from '../lib/error';
 import {
   StreamClient,
@@ -9,20 +9,16 @@ import {
   commonApi,
   ServiceError,
 } from '@viamrobotics/sdk';
-import { toast } from '../lib/toast';
-import InfoButton from './info-button.vue';
-import PCD from './pcd.vue';
-import { cameraStreamStates, baseStreamStates } from '../lib/camera-state';
+import { cameraStreamStates } from '../lib/camera-state';
 
 interface Props {
   cameraName: string;
+  parentName: string;
   resources: commonApi.ResourceName.AsObject[];
   client: Client;
-}
-
-interface Emits {
-  (event: 'clear-interval'): void;
-  (event: 'selected-camera-view', value: number): void;
+  showExportScreenshot: boolean;
+  refreshRate: string | undefined;
+  triggerRefresh: boolean;
 }
 
 const selectedMap = {
@@ -34,108 +30,116 @@ const selectedMap = {
 } as const;
 
 const props = defineProps<Props>();
-const emit = defineEmits<Emits>();
 
-let pcdExpanded = $ref(false);
-let pointcloud = $ref<Uint8Array | undefined>();
-let camera = $ref(false);
+const videoEl = $ref<HTMLVideoElement>();
+const imgEl = $ref<HTMLImageElement>();
 
-const selectedValue = $ref('Live');
+let cameraOn = $ref(false);
+let cameraFrameIntervalId = $ref(-1);
+let camerasOn = $ref(0);
 
-const initStreamState = () => {
-  cameraStreamStates.set(props.cameraName, false);
-};
-
-const clearStreamContainer = (camName: string, elName: string) => {
-  const streamContainer = document.querySelector(
-    `[data-stream="${camName}"]`
-  );
-  if (elName === 'video') {
-    streamContainer?.querySelector('video')?.classList.add('hidden');
-    streamContainer?.querySelector('img')?.classList.remove('hidden');
-  } else {
-    streamContainer?.querySelector('img')?.classList.add('hidden');
-    streamContainer?.querySelector('video')?.classList.remove('hidden');
+const manageStreamStates = () => {
+  let counter = 0;
+  for (const value of cameraStreamStates.values()) {
+    if (value.name === props.cameraName && value.on) {
+      counter++;
+    }
   }
+  camerasOn = counter;
 };
 
 const viewCamera = async (isOn: boolean) => {
   const streams = new StreamClient(props.client);
-  if (isOn) {
-    try {
-      // only add stream if not already active
-      if (
-        !baseStreamStates.get(props.cameraName) &&
-        !cameraStreamStates.get(props.cameraName)
-      ) {
+
+  streams.on('track', (event) => {
+    const eventStream = event.streams[0];
+    if (!eventStream) {
+      throw new Error('expected event stream to exist');
+    }
+    videoEl.srcObject = eventStream;
+  });
+
+  if (isOn && props.refreshRate === 'Live') {
+    cameraStreamStates.set(`${props.parentName}-${props.cameraName}`, {
+      on: true,
+      live: props.refreshRate === 'Live',
+      name: props.cameraName,
+    });
+    manageStreamStates();
+
+    if (camerasOn === 1) {
+      try {
         await streams.add(props.cameraName);
+      } catch (error) {
+        displayError(error as ServiceError);
       }
-    } catch (error) {
-      displayError(error as ServiceError);
     }
-    cameraStreamStates.set(props.cameraName, true);
-  } else {
-    try {
-      // only remove camera stream if active and base stream is not active
-      if (
-        !baseStreamStates.get(props.cameraName) &&
-        cameraStreamStates.get(props.cameraName)
-      ) {
+  } else if (props.refreshRate === 'Live') {
+    cameraStreamStates.set(`${props.parentName}-${props.cameraName}`, {
+      on: false,
+      live: props.refreshRate === 'Live',
+      name: props.cameraName,
+    });
+    manageStreamStates();
+
+    if (camerasOn === 0) {
+      try {
         await streams.remove(props.cameraName);
+      } catch (error) {
+        displayError(error as ServiceError);
       }
-    } catch (error) {
-      displayError(error as ServiceError);
     }
-    cameraStreamStates.set(props.cameraName, false);
   }
 };
 
-const renderPCD = async () => {
+const viewFrame = async (cameraName: string) => {
+  let blob;
   try {
-    pointcloud = await new CameraClient(props.client, props.cameraName).getPointCloud();
+    blob = await new CameraClient(props.client, cameraName).renderFrame(Camera.MimeType.JPEG);
   } catch (error) {
-    toast.error(`Error getting point cloud: ${error}`);
+    displayError(error as ServiceError);
+    return;
+  }
+
+  imgEl.setAttribute('src', URL.createObjectURL(blob));
+};
+
+const viewCameraFrame = (cameraName: string, time: number) => {
+  clearFrameInterval();
+
+  // Live
+  if (time === -1) {
+    return;
+  }
+
+  viewFrame(cameraName);
+  if (time > 0) {
+    cameraFrameIntervalId = window.setInterval(() => {
+      viewFrame(cameraName);
+    }, Number(time) * 1000);
   }
 };
 
-const togglePCDExpand = () => {
-  pcdExpanded = !pcdExpanded;
-  if (pcdExpanded) {
-    renderPCD();
-  }
+const clearFrameInterval = () => {
+  window.clearInterval(cameraFrameIntervalId);
 };
 
 const selectCameraView = () => {
-  emit('clear-interval');
+  clearFrameInterval();
+  const selectedInterval: number = selectedMap[props.refreshRate as keyof typeof selectedMap];
 
-  if (selectedValue !== 'Live') {
-    clearStreamContainer(props.cameraName, 'video');
-
-    const selectedInterval: number = selectedMap[selectedValue as keyof typeof selectedMap];
-    viewCamera(false);
-    emit('selected-camera-view', selectedInterval);
-    return;
-  }
-  clearStreamContainer(props.cameraName, 'img');
-
-  viewCamera(true);
-};
-
-const toggleExpand = () => {
-  camera = !camera;
-
-  emit('clear-interval');
-
-  if (camera) {
-    selectCameraView();
+  if (props.refreshRate === 'Live') {
+    videoEl.play();
   } else {
-    viewCamera(false);
+    viewCameraFrame(props.cameraName, selectedInterval);
   }
 };
 
 const refreshCamera = () => {
-  emit('selected-camera-view', selectedMap[selectedValue as keyof typeof selectedMap]);
-  emit('clear-interval');
+  const selectedInterval: number = selectedMap[props.refreshRate as keyof typeof selectedMap];
+
+  viewCameraFrame(props.cameraName, selectedInterval);
+  clearFrameInterval();
 };
 
 const exportScreenshot = async (cameraName: string) => {
@@ -153,100 +157,67 @@ const exportScreenshot = async (cameraName: string) => {
 };
 
 onMounted(() => {
-  initStreamState();
+  cameraOn = true;
+
+  clearFrameInterval();
+
+  selectCameraView();
+
+  if (props.refreshRate === 'Live') {
+    viewCamera(true);
+  }
 });
+
+onUnmounted(() => {
+  cameraOn = false;
+
+  clearFrameInterval();
+
+  viewCamera(false);
+});
+
+// on prop change select camera view
+watch(() => props.refreshRate, () => {
+  selectCameraView();
+});
+
+// on prop change refresh camera
+watch(() => props.triggerRefresh, () => {
+  refreshCamera();
+});
+
 </script>
 
 <template>
-  <v-collapse
-    :title="cameraName"
-    class="camera"
-  >
-    <v-breadcrumbs
-      slot="title"
-      crumbs="camera"
+  <div class="flex flex-col gap-4">
+    <template
+      v-if="cameraOn"
+    >
+      <v-button
+        v-if="cameraOn && props.showExportScreenshot"
+        icon="camera"
+        label="Export Screenshot"
+        @click="exportScreenshot(cameraName)"
+      />
+    </template>
+
+    <video
+      v-show="props.refreshRate === 'Live'"
+      ref="videoEl"
+      muted
+      autoplay
+      controls="false"
+      playsinline
+      :aria-label="`${cameraName} stream`"
+      :class="{ hidden: !cameraOn }"
+      class="clear-both h-fit transition-all duration-300 ease-in-out"
     />
-    <div class="h-auto border-x border-b border-black p-2">
-      <div class="container mx-auto">
-        <div class="pt-4">
-          <div class="flex mb-4 items-center gap-2">
-            <v-switch
-              id="camera"
-              :label="camera ? 'Hide Camera' : 'View Camera'"
-              :aria-label="
-                camera
-                  ? `Hide Camera: ${$props.cameraName}`
-                  : `View Camera: ${$props.cameraName}`
-              "
-              :value="camera ? 'on' : 'off'"
-              @input="toggleExpand"
-            />
-          </div>
 
-          <div class="pb-4">
-            <div class="flex flex-wrap">
-              <div
-                v-if="camera"
-                class="flex flex-wrap justify-items-end gap-2"
-              >
-                <div class="">
-                  <div class="relative">
-                    <v-select
-                      v-model="selectedValue"
-                      label="Refresh frequency"
-                      aria-label="Default select example"
-                      :options="Object.keys(selectedMap).join(',')"
-                      @input="selectCameraView"
-                    />
-                  </div>
-                </div>
-                <div class="self-end">
-                  <v-button
-                    v-if="camera && selectedValue === 'Manual Refresh'"
-                    icon="refresh"
-                    label="Refresh"
-                    @click="refreshCamera"
-                  />
-                </div>
-                <div class="self-end">
-                  <v-button
-                    v-if="camera"
-                    icon="camera"
-                    label="Export Screenshot"
-                    @click="exportScreenshot(cameraName)"
-                  />
-                </div>
-              </div>
-            </div>
-          </div>
-          <div
-            :aria-label="`${cameraName} stream`"
-            :data-stream="cameraName"
-            :class="{ hidden: !camera }"
-            class="clear-both h-fit transition-all duration-300 ease-in-out"
-          />
-        </div>
-        <div class="pt-4">
-          <div class="flex items-center gap-2 align-top">
-            <v-switch
-              :label="pcdExpanded ? 'Hide Point Cloud Data' : 'View Point Cloud Data'"
-              :value="pcdExpanded ? 'on' : 'off'"
-              @input="togglePCDExpand"
-            />
-            <InfoButton
-              :info-rows="['When turned on, point cloud will be recalculated']"
-            />
-          </div>
-
-          <PCD
-            v-if="pcdExpanded"
-            :resources="resources"
-            :pointcloud="pointcloud"
-            :camera-name="cameraName"
-            :client="client"
-          />
-        </div>
-      </div>
-    </div>
-  </v-collapse>
+    <img
+      v-show="props.refreshRate !== 'Live'"
+      ref="imgEl"
+      :aria-label="`${cameraName} stream`"
+      :class="{ hidden: props.refreshRate === 'Live' }"
+    >
+  </div>
 </template>
